refactor: socketIO 패키지 정리
