+++ toneup/src/main/java/com/threeboys/toneup/socketio/annotation/SocketController.java
@@ -1,14 +1,14 @@
-package com.threeboys.toneup.socketio;
+package com.threeboys.toneup.socketio.annotation;
+
+import org.springframework.web.bind.annotation.RestController;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface SocketMapping {
-    String endpoint();
-
-    Class<?> requestCls() default Void.class;
+@RestController
+public @interface SocketController {
 }
