+++ toneup/src/main/java/com/threeboys/toneup/socketio/annotation/SocketMapping.java
@@ -1,14 +1,14 @@
-package com.threeboys.toneup.socketio;
-
-import org.springframework.web.bind.annotation.RestController;
+package com.threeboys.toneup.socketio.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.TYPE)
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-@RestController
-public @interface SocketController {
+public @interface SocketMapping {
+    String endpoint();
+
+    Class<?> requestCls() default Void.class;
 }
